Add JSON format option to ipam-whatis and ipam-select
These commands now accept the argument --json=compact|pretty to
produce output in JSON.  The "compact" argument generates a single
line of output while the "pretty" argument genrates nicely formatted
output.

The bulk of the work has been integrated into the IPAM::nameinfo() and
IPAM::prefixinfo() methods, which return a Perl hash that can be
translated to JSON as is.  ipam-whatis and ipam-select have been
rewritten to use these methods and provide the JSON format as well as
the traditional output in prose.

A performance bottleneck has been eliminated by creating a new
top-level registry IPAM::REG_ALIAS in the IPAM object, which holds
lists of Host objects that provide a canonical name for all known
aliases.