Support zones on remote master servers
The Makefile mechanics recognizes the convention to specify a remote
DNS server in the name of a zone's directory as "zone_dir@server" and
invokes an external command to push the zone file snippet to the
server rather than copying it to a local repository.

To allow this, the new() method of Zone.pm no longer checks for the
existence of the zone's directory.